+++ resources/js/components/app-sidebar.tsx
@@ -4,7 +4,7 @@ import { Sidebar, SidebarContent, SidebarFooter, SidebarHeader, SidebarMenu, Sid
 import { type NavItem } from '@/types';
 import { Link } from '@inertiajs/react';
 import { Calendar, GraduationCap, BookOpen as Journal, LayoutGrid, Newspaper, Users, Video, UserCircle, UserRoundPen, Building } from 'lucide-react';
-import AppLogo from './app-logo';
+import AppLogo from '@/components/atoms/app-logo';
 
 const mainNavItems: NavItem[] = [
     {
