+++ resources/js/components/atoms/app-logo.tsx
@@ -1,5 +1,3 @@
-import appLogo from '../../../public/logoascee.png'
-
 interface AppLogoProps {
     className?: string;
 }
@@ -8,7 +6,7 @@ export default function AppLogo({ className }: AppLogoProps) {
     return (
         <>
             <div className="flex items-center w-20 h-auto rounded-full ">
-                <img src={appLogo} className={className} alt="ASCEE Logo"/>
+                <img src="/logoascee.png" className={className} alt="ASCEE Logo"/>
             </div>
         </>
     );
