+++ resources/js/layouts/settings/layout.tsx
@@ -1,4 +1,4 @@
-import Heading from '@/components/heading';
+import Heading from '@/components/atoms/heading';
 import { Button } from '@/components/ui/button';
 import { Separator } from '@/components/ui/separator';
 import { cn } from '@/lib/utils';
